refactor: deduplicate control character detection into shared validation utility
