+++ config-schema.json
@@ -50,7 +50,16 @@
           "type": "array",
           "items": {
             "type": "string",
-            "enum": ["rules", "ignore", "mcp", "subagents", "commands", "skills", "hooks", "*"]
+            "enum": [
+              "rules",
+              "ignore",
+              "mcp",
+              "subagents",
+              "commands",
+              "skills",
+              "hooks",
+              "*"
+            ]
           }
         },
         {
@@ -62,7 +71,16 @@
             "type": "array",
             "items": {
               "type": "string",
-              "enum": ["rules", "ignore", "mcp", "subagents", "commands", "skills", "hooks", "*"]
+              "enum": [
+                "rules",
+                "ignore",
+                "mcp",
+                "subagents",
+                "commands",
+                "skills",
+                "hooks",
+                "*"
+              ]
             }
           }
         }
@@ -112,7 +130,10 @@
           },
           "transport": {
             "type": "string",
-            "enum": ["github", "git"]
+            "enum": [
+              "github",
+              "git"
+            ]
           },
           "ref": {
             "type": "string"
@@ -121,7 +142,9 @@
             "type": "string"
           }
         },
-        "required": ["source"],
+        "required": [
+          "source"
+        ],
         "additionalProperties": false
       }
     }
